Warn if KUBECONFIG is not set
Add a warning if the KUBECONFIG environment variable is not set. This is
important if developers want/need to provide their own cluster
credentials, and not rely on a default kubeconfig file added to the
Jenkins build agent environment. Testing is accomplished with the
Kubernetes CLI plugin, which provides the `withKubeConfig` wrapper for
Pipeline DSL syntax and injects a KUBECONFIG env var.