Cleanups for MVP Release Candidate
This change gets rid of many items that were added as part of the
"Hello World" plugin archetype, such as the TknAction class and its
resources, plus the unused "Name" property and its translations.
The TknBuilder form was updated to have its property titles match the
respective code property. Help text was also updated (and added for
the "Commands" textbox), and the plugin now has a proper description
for the various "plugins" management screens in the Jenkins web console.

The README was also updated to provide MVP-worthy documentation.
Includes instructions on how to set up the Tkn Client Plugin and use it
in a Jenkins pipeline or a Freestyle build. Screen shots were added to
improve the developer experience. Finally, a link to the Tekton Client
Plugin was added for proper attribution and to clarify potential
confusion.