Demo - Pipelines as Code Squared
This demos the tkn-client-plugin on OpenShift, using the OpenShift
Pipelines operator and the Jenkins ephemeral template. The demo
utilizes an extension of the OpenShift Jenkins image that adds the
tkn client plugin (built locally), and has a README with full
instructions on how to deploy the demo yourself. The demo is called
"Pipelines as Code - Squared" because it uses Jenkins as Code (from
a Jenkinsfile) to orchestrate Tekton as Code (from a Pipeline YAML
manifest).

To get this to work on OpenShift 4.13, the minimum supported version
of Jenkins was downgraded to 2.401.1.